requirements.txt
Django==4.2.16
aiogram==3.15.0
python-dotenv==1.0.1